+++ driver-service/src/main/kotlin/com/example/driver_service/model/entity/DriverEntity.kt
@@ -2,6 +2,7 @@ package com.example.driver_service.model.entity
 
 import com.example.driver_service.model.enums.Gender
 import com.example.driver_service.model.enums.WorkStatus
+import java.math.BigDecimal
 import java.time.LocalDateTime
 import java.util.UUID
 
@@ -12,7 +13,7 @@ data class DriverEntity(
     var email: String = "",
     var password: String = "",
     var phoneNumber: String = "",
-    var rating: Float? = null,
+    var rating: BigDecimal? = null,
     var gender: Gender = Gender.OTHER,
     var carId: UUID? = null,
     var workStatus: WorkStatus = WorkStatus.OFF_DUTY,
